+++ examples/payment.rs
@@ -44,6 +44,8 @@ async fn main() -> Result<(), bepaid::BepaidError> {
                 device_id: None,
                 birth_date: None,
                 phone: None,
+                external_id: None,
+                taxpayer_id: None,
             }),
             billing_address: None,
             custom_fields: None,